init: add dependencies and customization
