+++ .gitignore
@@ -14,3 +14,6 @@ Thumbs.db
 
 # Logs
 npm-debug.log
+
+# Folder bild
+dist
